Implement go wrapper for ebpf_obj_pin(), refactor map and program_base accordingly.
